fix: convert opportunity.amount (common schema) to float

+++ packages/core/destination_writers/bigquery.ts
@@ -911,7 +911,7 @@ const schemaByCommonObjectType: {
         },
         {
           name: 'amount',
-          type: 'INTEGER',
+          type: 'FLOAT64',
           mode: 'NULLABLE',
         },
         {
